created index and directory for recipes
Created and initialized the index.html file with the boilerplate and added links to two recipes.
Additionally, a separate directory for the recipes was made, along with their respective html files.
Links to the mentioned pages were added to the index.html file. The README file was editted to include
a description of the project.